Avant Login

+++ src/Controller/ContientController.php
@@ -44,7 +44,8 @@ class ContientController extends AbstractFOSRestController
      */
     public function postCommandeContientAction(Commande $commande, ParamFetcher $paramFetcher)
     {
-        $contient = new Contient();
+        if(!($commande -> getLivree())){
+            $contient = new Contient();
         $meuble = $this -> getDoctrine() -> getRepository(Meuble::class) -> find($paramFetcher -> get("numserie"));
 
         $nombreCommandee = $paramFetcher -> get("nombrecommande");
@@ -67,6 +68,10 @@ class ContientController extends AbstractFOSRestController
         }else{
             return $this -> view(['messages' => 'Erreur de commande'], Response::HTTP_NOT_ACCEPTABLE) ;
         }
+        }else{
+            return $this -> view(['messages' => 'Erreur ajout commande livree'], Response::HTTP_NOT_ACCEPTABLE) ;
+        }
+        
     }
 
     /**
@@ -75,49 +80,65 @@ class ContientController extends AbstractFOSRestController
      */
     public function putCommandeContientAction(Commande $commandeA, Meuble $meubleA, ParamFetcher $paramFetcher)
     {
-        $contient = $this -> contientRepository -> findOneBy(['commandeNumCommande' => $commandeA, 'meubleNumSerie' => $meubleA]);
+        if(!($commandeA -> getLivree())){
 
-        $meuble = $this -> getDoctrine() -> getRepository(Meuble::class) -> find($paramFetcher -> get("numserie"));
-
-        $nombreStock = $meuble -> getQuantiteStock();
-        $nombreCommandee = $paramFetcher -> get("nombrecommande");
-
-        if($nombreCommandee >= $nombreStock){
-            $newStock = $nombreStock - $nombreCommandee;
-            $contient -> setMeubleNumSerie($meuble)
-            -> setNombreCommande()
-            ;
-            $meuble -> setQuantiteStock($newStock);
-
-            $this -> entityManager -> persist($contient);
-            $this -> entityManager -> persist($meuble);
-            $this -> entityManager -> flush();
-            return $this -> view($contient, Response::HTTP_OK);
+            $contient = $this -> contientRepository -> findOneBy(['commandeNumCommande' => $commandeA, 'meubleNumSerie' => $meubleA]);
+    
+            $meuble = $this -> getDoctrine() -> getRepository(Meuble::class) -> find($paramFetcher -> get("numserie"));
+    
+            $nombreStock = $meuble -> getQuantiteStock();
+            $nombreCommandee = $paramFetcher -> get("nombrecommande");
+    
+            if($nombreCommandee >= $nombreStock){
+                $newStock = $nombreStock - $nombreCommandee;
+                $contient -> setMeubleNumSerie($meuble)
+                -> setNombreCommande()
+                ;
+                $meuble -> setQuantiteStock($newStock);
+    
+                $this -> entityManager -> persist($contient);
+                $this -> entityManager -> persist($meuble);
+                $this -> entityManager -> flush();
+                return $this -> view($contient, Response::HTTP_OK);
+            }else{
+                return $this -> view(['messages' => 'Erreur de commande'], Response::HTTP_NOT_ACCEPTABLE) ;
+            }
+        }else{
+            return $this -> view(['messages' => 'Erreur ajout commande livree'], Response::HTTP_NOT_ACCEPTABLE) ;
         }
     }
 
-    
-
-    public function deleteCommandeContientsAction(Commande $commandeA, Meuble $meubleA)
+    public function deleteCommandeContientsAction(Commande $commande, Meuble $meuble)
     {   
-        $contient = $this -> contientRepository -> findOneBy(['commandeNumCommande' => $commandeA, 'meubleNumSerie' => $meubleA]);
+        if(!($commande -> getLivree())){
+            $contient = $this -> contientRepository -> findOneBy(['commandeNumCommande' => $commande, 'meubleNumSerie' => $meuble]);
+
+            $nombreCommandee = $contient -> getNombreCommande();
+            $nombreStock = $meuble -> getQuantiteStock();
+            $oldCommande = $nombreCommandee + $nombreStock;
+            $meuble -> setQuantiteStock($oldCommande);
             $this -> entityManager -> remove($contient);
             $this -> entityManager -> flush();
             return $this -> view($contient, Response::HTTP_OK);
+        }else{
+            return $this -> view(['messages' => 'Erreur suppression commande livree'], Response::HTTP_NOT_ACCEPTABLE) ;
+        }
     }
-
-    public function deleteCommandeContientsCancelAction(Commande $commande, Meuble $meuble)
-    {   
-        $contient = $this -> contientRepository -> findOneBy(['commandeNumCommande' => $commande, 'meubleNumSerie' => $meuble]);
-
-        $nombreCommandee = $contient -> getNombreCommande();
-        $nombreStock = $meuble -> getQuantiteStock();
-        $oldCommande = $nombreCommandee + $nombreStock;
-        $meuble -> setQuantiteStock($oldCommande);
-        $this -> entityManager -> remove($contient);
-        $this -> entityManager -> flush();
-        return $this -> view($contient, Response::HTTP_OK);
-    }
-
     // FIN CRUD Contients
+
+    //CRUD facture
+
+    public function getFactureAction(Commande $commande)
+    {
+        $manager = $this->getDoctrine()->getManager();
+        $query = "SELECT nom_Client as nomClient, prenom_Client as prenomClient, tel_client as telClient, client.adresse_client as adresseClient ,CONCAT(client.province_client, ' ', client.cp_client) as provinceClient, commande.num_commande as numCommande, commande.date_commande as dateCommande, contient.meuble_num_serie as numSerie, meuble.nom_meuble as nomMeuble, contient.nombre_commande as quantiteCommande, meuble.prix as prixUnitaire, (meuble.prix * contient.nombre_commande) as prixTotal, meuble.categorie from client inner join commande on client.num_client = commande.client_num_client inner join contient on commande.num_commande = contient.commande_num_commande inner join meuble on contient.meuble_num_serie = meuble.num_serie WHERE commande.num_commande = :numCommande";
+        $statement = $manager->getConnection()->prepare($query);
+        $statement->execute(['numCommande' => $commande -> getNumCommande()]);
+        $facture = $statement->fetchAll();
+        return $this -> view($facture, Response::HTTP_OK) ;
+    }
+
+
+    //fin crud facture
+
 }
